+++ common/tools/dev-tool/src/commands/run/testNodeTSInput.ts
@@ -17,7 +17,7 @@ export const commandInfo = makeCommandInfo(
       default: false,
       description: "whether to disable launching test-proxy",
     },
-  }
+  },
 );
 
 export default leafCommand(commandInfo, async (options) => {
@@ -26,7 +26,7 @@ export default leafCommand(commandInfo, async (options) => {
     isModuleProj ? "" : "-r esm "
   }-r ts-node/register --reporter ../../../common/tools/mocha-multi-reporter.js --full-trace`;
   const updatedArgs = options["--"]?.map((opt) =>
-    opt.includes("**") && !opt.startsWith("'") && !opt.startsWith('"') ? `"${opt}"` : opt
+    opt.includes("**") && !opt.startsWith("'") && !opt.startsWith('"') ? `"${opt}"` : opt,
   );
   const mochaArgs = updatedArgs?.length
     ? updatedArgs?.join(" ")