+++ sdk/appservice/arm-appservice/samples-dev/webAppsRestoreSample.ts
@@ -18,7 +18,7 @@ dotenv.config();
  * This sample demonstrates how to Description for Restores a specific backup to another app (or deployment slot, if specified).
  *
  * @summary Description for Restores a specific backup to another app (or deployment slot, if specified).
- * x-ms-original-file: specification/web/resource-manager/Microsoft.Web/stable/2022-09-01/examples/RestoreWebAppBackup.json
+ * x-ms-original-file: specification/web/resource-manager/Microsoft.Web/stable/2023-01-01/examples/RestoreWebAppBackup.json
  */
 async function restoreWebAppFromBackup() {
   const subscriptionId =