+++ sdk/appservice/arm-appservice/samples/v14/javascript/staticSitesGetBuildDatabaseConnectionsWithDetailsSample.js
@@ -16,7 +16,7 @@ require("dotenv").config();
  * This sample demonstrates how to Returns details of database connections for a static site build
  *
  * @summary Returns details of database connections for a static site build
- * x-ms-original-file: specification/web/resource-manager/Microsoft.Web/stable/2022-09-01/examples/GetStaticSiteBuildDatabaseConnectionsWithDetails.json
+ * x-ms-original-file: specification/web/resource-manager/Microsoft.Web/stable/2023-01-01/examples/GetStaticSiteBuildDatabaseConnectionsWithDetails.json
  */
 async function listFullDetailsOfDatabaseConnectionsForTheStaticSiteBuild() {
   const subscriptionId =