+++ sdk/appservice/arm-appservice/samples/v14/typescript/src/appServiceEnvironmentsGetSample.ts
@@ -18,7 +18,7 @@ dotenv.config();
  * This sample demonstrates how to Description for Get the properties of an App Service Environment.
  *
  * @summary Description for Get the properties of an App Service Environment.
- * x-ms-original-file: specification/web/resource-manager/Microsoft.Web/stable/2022-09-01/examples/AppServiceEnvironments_Get.json
+ * x-ms-original-file: specification/web/resource-manager/Microsoft.Web/stable/2023-01-01/examples/AppServiceEnvironments_Get.json
  */
 async function getThePropertiesOfAnAppServiceEnvironment() {
   const subscriptionId =