+++ sdk/hybridcontainerservice/arm-hybridcontainerservice/samples/v1-beta/typescript/src/hybridIdentityMetadataListByClusterSample.ts
@@ -15,24 +15,19 @@ import * as dotenv from "dotenv";
 dotenv.config();
 
 /**
- * This sample demonstrates how to Lists the hybrid identity metadata proxy resource in a cluster.
+ * This sample demonstrates how to Lists the hybrid identity metadata proxy resource in a provisioned cluster instance.
  *
- * @summary Lists the hybrid identity metadata proxy resource in a cluster.
- * x-ms-original-file: specification/hybridaks/resource-manager/Microsoft.HybridContainerService/preview/2022-09-01-preview/examples/HybridIdentityMetadataListByCluster.json
+ * @summary Lists the hybrid identity metadata proxy resource in a provisioned cluster instance.
+ * x-ms-original-file: specification/hybridaks/resource-manager/Microsoft.HybridContainerService/preview/2023-11-15-preview/examples/HybridIdentityMetadataListByCluster.json
  */
 async function hybridIdentityMetadataListByCluster() {
-  const subscriptionId =
-    process.env["HYBRIDCONTAINERSERVICE_SUBSCRIPTION_ID"] ||
-    "fd3c3665-1729-4b7b-9a38-238e83b0f98b";
-  const resourceGroupName =
-    process.env["HYBRIDCONTAINERSERVICE_RESOURCE_GROUP"] || "testrg";
-  const resourceName = "ContosoTargetCluster";
+  const connectedClusterResourceUri =
+    "subscriptions/fd3c3665-1729-4b7b-9a38-238e83b0f98b/resourceGroups/testrg/providers/Microsoft.Kubernetes/connectedClusters/test-hybridakscluster";
   const credential = new DefaultAzureCredential();
-  const client = new HybridContainerServiceClient(credential, subscriptionId);
+  const client = new HybridContainerServiceClient(credential);
   const resArray = new Array();
   for await (let item of client.hybridIdentityMetadataOperations.listByCluster(
-    resourceGroupName,
-    resourceName
+    connectedClusterResourceUri
   )) {
     resArray.push(item);
   }